% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MNPprocessIDAT_functions.R
\name{MNPpreprocessIllumina}
\alias{MNPpreprocessIllumina}
\alias{MNPnormalize.illumina.control}
\title{Modified version of the \link[minfi]{preprocessIllumnia} normalization function}
\usage{
MNPpreprocessIllumina(
  rgSet,
  bg.correct = TRUE,
  normalize = c("controls", "no"),
  ref = 10000
)

MNPnormalize.illumina.control(rgSet, ref = 10000)
}
\arguments{
\item{rgSet}{an object of class RGset}

\item{bg.correct}{background correction}

\item{normalize}{either "controls" or "no"}

\item{ref}{reference intensity 10,000}

\item{RGset}{object of class RGset}
}
\description{
\code{MNPpreprocessIllumina} works with single samples. Normalize to 10,000 instead of average normalization control intensity of first/reference sample
}
\details{
See \link[minfi]{preprocessIllumnia} for more details.
}
\section{Functions}{
\itemize{
\item \code{MNPnormalize.illumina.control}: Modified version of \code{normalize.illumina.control} called by\code{\link{MNPpreprocessIllumina}}

\code{MNPnormalize.illumina.control} works with single samples. Normalize to 10,000 instead of average normalization control intensity of first/reference sample
Called by \code{\link{MNPpreprocessIllumina}}
}}

